Fit an illness-death model using sampling or boostrap

Fit an illness-death model using sampling or boostrap

bootstrap.penidm(
  K = 100,
  tau = 0.5,
  seed = 1,
  resampling = "subsampling",
  formula01,
  formula02,
  formula12,
  data,
  method = "Weib",
  scale.X = T,
  maxiter = 100,
  maxiter.pena = 10,
  eps = c(5, 5, 3),
  n.knots = NULL,
  knots = "equidistant",
  type.quantile = 1,
  na.action = na.fail,
  B = NULL,
  posfix = NULL,
  gausspoint = 10,
  lambda01 = NULL,
  lambda02 = NULL,
  lambda12 = NULL,
  nlambda01 = 50,
  nlambda02 = 50,
  nlambda12 = 50,
  penalty = "lasso",
  penalty.factor = NULL,
  alpha = ifelse(penalty == "scad", 3.7, ifelse(penalty == "mcp", 3, ifelse(penalty
    %in% c("elasticnet", "corrected.elasticnet"), 0.5, 1))),
  step.sequential = F,
  option.sequential = list(cutoff = 10^-3, min = 20, step = 10),
  nproc = 1,
  clustertype = "FORK"
)

Arguments

K

number of repetitions of sampling or boostrap, by default values 100

tau

Percentage of observation to use for resampling can go u

seed

value of the seed to initialise the random number generator and ensure reproducibility of the results

resampling

approach. Possible values are: "subsampling" for sampling without replacement of a proportion tau of the observations, or "bootstrap" for sampling with replacement generating a resampled dataset with as many observations as in the full sample. Alternatively, this argument can be a function to use for resampling. This function must use arguments named data and tau and return the IDs of observations to be included in the resampled dataset.

formula01

A formula specifying a regression model for the 0 --> 1 transition from the initial state to the transient state of the illness-death model. The right hand side of the formula specifies the covariate terms, and the left hand side must be an event history object as returned by the function Hist.

formula02

A formula specifying a regression model for the 0 --> 2 transition from the initial state to the absorbing state. The left hand side must be equal to the left hand side of formula01. If missing it is set to formula01.

formula12

A formula specifying a regression model for the 1 --> 2 transition from the transient state to the absorbing state. operator is not required. If missing it is set to formula01.

data

A data frame in which to interpret the variables of formula01, formula02 and formula12.

method

type of estimation method: "splines" for a likelihood approach with approximation of the transition intensities by M-splines, "Weib" for a parametric approach with a Weibull distribution on the transition intensities. Default is "Weib".

scale.X

do you want to center and reduce your explanatory variables

maxiter

Maximum number of iterations. The default is 200.

maxiter.pena

Maximum number of iterations for penalised coefficients

eps

A vector of 3 integers >0 used to define the power of three convergence criteria: 1. for the regression parameters, 2. for the likelihood, 3. for the second derivatives. The default is c(5,5,3) which is translated into convergence if the respective values change less then \(10^{-5}\) (for regression parameters and likelihood) and \(10^{-3}\) for the second derivatives between two iterations.

n.knots

For method="splines" only, a vector of length 3 specifing the number of knots, one for each transition, for the M-splines estimate of the baseline intensities in the order 0 --> 1, 0 --> 2, 1 --> 2. The default is c(7,7,7). When knots are specified as a list this argument is ignored. The algorithm needs least 5 knots and at most 20 knots.

knots

Argument only active for the penalized likelihood approach method="Splines". There are three ways to control the placement of the knots between the smallest and the largest of all time points:

knots="equidistant"

Knots are placed with same distance on the time scale.

knots="quantile"

Knots are placed such that the number of observations is roughly the same between knots.

knots=list()

List of 1 or 2 or three vectors. The list elements are the actual placements (timepoints) of the knots for the M-spline. The list may contain one vector of placements for each transition in the order 0 --> 1, 0 --> 2, 1 --> 2. If only vector is specifified the knots are used for all transitions. If only 2 vectors are specifified, the knots for the 0 --> 1 transition are also used for the 1 --> 2 transition.

The algorithm needs at least 3 knots in spline and allows no more than 20 knots.

type.quantile

Argument only active for the likelihood approach method="splines". There are three ways to control the placement of the knots according to the time considered between states :=

#'
type.quantile=1

Time for 0 --> 1 is the imputed to the middle of the interval left and right for demence . Time for 0 --> 2 and 1 --> 2 is the same t, time of news.

type.quantile=2

Time for 0 --> 1 is the imputed to the middle of the interval left and right. Time for 0 --> 2 and 1 --> 2 is the same t, time of news.

type.quantile=3

Time for 0 --> 1 is the imputed to the middle of the interval left and right. Time for 0 --> 2 is time of death for non demented sujects only. Time for 1 --> 2 is time of death for suject diagnose with dementia.

#'
type.quantile=4

Time for 0 --> 1 is left and right. Time for 0 --> 2 is time of death for non demented sujects only. Time for 1 --> 2 is time of death for suject diagnose with dementia.

na.action

how NAs are treated. The default is first, any na.action attribute of data, second a na.action setting of options, and third 'na.fail' if that is unset. The 'factory-fresh' default is na.omit. Another possible value is NULL.

B

vector of size the number of parameters, in the following order, first the parameters of splines 0 --> 1, 0 --> 2, 1 --> 2, second the parameters of explanatory variables in order 0 --> 1, 0 --> 2, 1 --> 2. This argument is only used for models with M-splines.

posfix

index of fixed parameters

gausspoint

gauss quadrature points in the approximation of integrals

lambda01

Lambda on transition 0 –> 1

lambda02

Lambda on transition 0 –> 2

lambda12

Lambda on transition 1 –> 2

nlambda01

number of Lambda on transition 0 –> 1

nlambda02

number of Lambda on transition 0 –> 2

nlambda12

number of Lambda on transition 1 –> 2

penalty

which penalty to consider

penalty.factor

which variable should be penalised

alpha

alpha on all transitions

step.sequential

should we use the optimisation version to fix splines

option.sequential

parameters to give if you want to do the optimisation version to fix splines

nproc

number of cluster

clustertype

in which cluster to work

Value

The boostrap returns a list of K elements of type idm, see idm for more details

See also

Author

R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>

Examples


if (FALSE) { # \dontrun{
 library(lava)
 library(prodlim)
 set.seed(17)
 d <- simulateIDM(n=1000)$data
 fitweib <- bootstrap.idm(formula01=Hist(time=list(L,R),event=seen.ill)~X1+X2,
 formula02=Hist(time=observed.lifetime,event=seen.exit)~X1+X2,
 formula12=Hist(time=observed.lifetime,event=seen.exit)~X1+X2,data=d,penalty="none",K=3)
 print(fitweib)

} # }