First and second order derivatives of the log-likelihood with weibull baseline risk

First and second order derivatives of the log-likelihood with weibull baseline risk

derivaweib(
  b,
  npm,
  npar,
  bfix,
  fix,
  ctime,
  no,
  ve01,
  ve02,
  ve12,
  dimnva01,
  dimnva02,
  dimnva12,
  nva01,
  nva02,
  nva12,
  t0,
  t1,
  t2,
  t3,
  troncature
)

Arguments

b

parameters on explanatory variables not fixed

npm

number of parameters not fixed, thus length of b

npar

number of parameters

bfix

parameters on explanatory variables fixed

fix

indicator of length npar, values 1 if parameter fixed

ctime

profile of patients from 1 to 7

no

number of subjects

ve01

variables for transition 0 –>1

ve02

variables for transition 0 –>2

ve12

variables for transition 1 –>2

dimnva01

number of variables for transition 0 –>1, if not variables value 1

dimnva02

number of variables for transition 0 –>2, if not variables value 1

dimnva12

number of variables for transition 1 –>2, if not variables value 1

nva01

number of variables for transition 0 –>1, if not variables value 0

nva02

number of variables for transition 0 –>2, if not variables value 0

nva12

number of variables for transition 1 –>2, if not variables value 0

t0

time of entry

t1

time of last visit or last visit without diagnose of illness

t2

time of last visit or time diagnose of illness

t3

time of last visit or death

troncature

indicator of troncature, value 1 if there is troncature otherwise 0.

Author

R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>