Fit an illness-death model with limited and numerous explanatory variables

Fit an illness-death model with a limited number of explanatory variables using semi-parametric approach with a linear combination of M-splines or Weibull distribution on the transition intensities. Fit an illness-death model with a large number of explanatory variables using regularization approach of the log-likelihood, i.e. lasso, elastic-net, MCP or SCAD penalty. Left-truncated, right-censored, and interval-censored data are allowed. State 0 corresponds to the initial state, state 1 to the transient one, state 2 to the absorbant one. The allowed transitions are: 0 –> 1, 0 –> 2 and 1 –> 2.

The estimated parameters are obtained using the robust Marquardt algorithm (Marquardt, 1963) or a proximal gradient algorithm.

idm(
  formula01,
  formula02,
  formula12,
  data,
  method = "Weib",
  scale.X = T,
  maxiter = 100,
  maxiter.pena = 10,
  eps = c(5, 5, 3),
  n.knots = NULL,
  knots = "equidistant",
  type.quantile = 1,
  na.action = na.fail,
  B = NULL,
  posfix = NULL,
  gausspoint = 10,
  timedep12 = FALSE,
  semiMarkov = TRUE,
  step.sequential = F,
  option.sequential = list(cutoff = 10^-3, min = 20, step = 10),
  lambda01 = NULL,
  lambda02 = NULL,
  lambda12 = NULL,
  nlambda01 = 50,
  nlambda02 = 50,
  nlambda12 = 50,
  penalty = NULL,
  penalty.factor = NULL,
  alpha = ifelse(penalty == "scad", 3.7, ifelse(penalty == "mcp", 3, ifelse(penalty
    %in% c("elasticnet"), 0.5, 1))),
  nproc = 1,
  analytics = T,
  clustertype = "FORK",
  envir = parent.frame()
)

Arguments

formula01

A formula specifying a regression model for the 0 --> 1 transition from the initial state to the transient state of the illness-death model. The right hand side of the formula specifies the covariate terms, and the left hand side must be an event history object as returned by the function Hist.

formula02

A formula specifying a regression model for the 0 --> 2 transition from the initial state to the absorbing state. The left hand side must be equal to the left hand side of formula01. If missing it is set to formula01.

formula12

A formula specifying a regression model for the 1 --> 2 transition from the transient state to the absorbing state. operator is not required. If missing it is set to formula01.

data

A data frame in which to interpret the variables of formula01, formula02 and formula12.

method

type of estimation method: "splines" for a likelihood approach with approximation of the transition intensities by M-splines, "Weib" for a parametric approach with a Weibull distribution on the transition intensities. Default is "Weib".

scale.X

do you want to center and reduce your explanatory variables

maxiter

Maximum number of iterations. The default is 200.

maxiter.pena

Maximum number of iterations for penalised coefficients

eps

A vector of 3 integers >0 used to define the power of three convergence criteria: 1. for the regression parameters, 2. for the likelihood, 3. for the second derivatives. The default is c(5,5,3) which is translated into convergence if the respective values change less then \(10^{-5}\) (for regression parameters and likelihood) and \(10^{-3}\) for the second derivatives between two iterations.

n.knots

For method="splines" only, a vector of length 3 specifing the number of knots, one for each transition, for the M-splines estimate of the baseline intensities in the order 0 --> 1, 0 --> 2, 1 --> 2. The default is c(7,7,7). When knots are specified as a list this argument is ignored. The algorithm needs least 5 knots and at most 20 knots.

knots

Argument only active for the penalized likelihood approach method="Splines". There are three ways to control the placement of the knots between the smallest and the largest of all time points:

knots="equidistant"

Knots are placed with same distance on the time scale.

knots="quantile"

Knots are placed such that the number of observations is roughly the same between knots.

knots=list()

List of 1 or 2 or three vectors. The list elements are the actual placements (timepoints) of the knots for the M-spline. The list may contain one vector of placements for each transition in the order 0 --> 1, 0 --> 2, 1 --> 2. If only vector is specifified the knots are used for all transitions. If only 2 vectors are specifified, the knots for the 0 --> 1 transition are also used for the 1 --> 2 transition.

The algorithm needs at least 3 knots in spline and allows no more than 20 knots.

type.quantile

Argument only active for the likelihood approach method="splines". There are three ways to control the placement of the knots according to the time considered between states :=

type.quantile=1

Time for 0 --> 1 is the imputed to the middle of the interval left and right for demence . Time for 0 --> 2 and 1 --> 2 is the same t, time of news.

type.quantile=2

Time for 0 --> 1 is the imputed to the middle of the interval left and right. Time for 0 --> 2 and 1 --> 2 is the same t, time of news.

type.quantile=3

Time for 0 --> 1 is the imputed to the middle of the interval left and right. Time for 0 --> 2 is time of death for non demented sujects only. Time for 1 --> 2 is time of death for suject diagnose with dementia.

type.quantile=4

Time for 0 --> 1 is left and right. Time for 0 --> 2 is time of death for non demented sujects only. Time for 1 --> 2 is time of death for suject diagnose with dementia.

na.action

how NAs are treated. The default is first, any na.action attribute of data, second a na.action setting of options, and third 'na.fail' if that is unset. The 'factory-fresh' default is na.omit. Another possible value is NULL.

B

vector of size the number of parameters, in the following order, first the parameters of splines 0 --> 1, 0 --> 2, 1 --> 2, second the parameters of explanatory variables in order 0 --> 1, 0 --> 2, 1 --> 2. This argument is only used for models with M-splines.

posfix

index of fixed parameters

gausspoint

gauss quadrature points in the approximation of integrals in the likelihood

timedep12

TRUE if time dependent on 1 –> 2 otherwise FALSE (default)

semiMarkov

TRUE if semi Markov on 1 –> 2 otherwise FALSE (default)

step.sequential

should we use the optimisation version to fix splines

option.sequential

parameters to give if you want to do the optimisation version to fix splines

lambda01

Lambda on transition 0 –> 1

lambda02

Lambda on transition 0 –> 2

lambda12

Lambda on transition 1 –> 2

nlambda01

number of Lambda on transition 0 –> 1

nlambda02

number of Lambda on transition 0 –> 2

nlambda12

number of Lambda on transition 1 –> 2

penalty

which penalty to consider

penalty.factor

which variable should be penalised complete hessian is computed.

alpha

alpha on all transitions

nproc

number of cluster

analytics

TRUE if analytical hessian for M-spline otherwise FALSE - diff finite (default)

clustertype

in which cluster to work

envir

working environment

Value

call

the call that produced the result.

coef

regression parameters.

loglik

vector containing the log-likelihood and the penalised log-likelihood

cv

vector containing the convergence criteria based on stability of parameters, log-likelihood (or penalised) and relative distance to minimum/maximum

niter

number of iterations.

converged

integer equal to 1 when the model converged, 2, 3 or 4 otherwise.

modelPar

Weibull parameters.

N

number of subjects.

events1

number of events 0 –> 1.

events2

number of events 0 –> 2 or 0 –> 1 –> 2.

NC

vector containing the number of covariates on transitions 0 –> 1, 0 –> 2, 1 –> 2.

responseTrans

model response for the 0 –> 1 transition. Hist or Surv object.

responseAbs

model response for the 0 –> 2 transition. Hist or Surv object.

time

times for which transition intensities have been evaluated for plotting.

maxtime

times of last follow-up or event

mintime

times of first follow-up or event

HR

vector of hazard risks.

V

variance-covariance matrix derived from the Hessian of the log-likelihood or penalised log-likelihood

Xnames01

names of covariates on 0 –> 1.

Xnames02

names of covariates on 0 –> 2.

Xnames12

names of covariates on 1 –> 2.

knots01

knots to approximate by M-splines the intensity of the 0 –> 1 transition.

knots02

knots to approximate by M-splines the intensity of the 0 –> 2 transition.

knots12

knots to approximate by M-splines the intensity of the 1 –> 2 transition.

nknots01

number of knots on transition 0 –> 1.

nknots02

number of knots on transition 0 –> 2.

nknots12

number of knots on transition 1 –> 2.

theta01

square root of splines coefficients for transition 0 –> 1.

theta02

square root of splines coefficients for transition 0 –> 2.

theta12

square root of splines coefficients for transition 1 –> 2.

alpha

penalty parameter alpha on all transitions

lambda

matrix of lambda penalty parameters, first line for 0 –> 1, second line for 0 –> 2 and third line for 1 –> 2

BIC

Bayesian Information Criterion

GCV

Generalised Cross-Validation approximation

levels

a list containing the type of the variable on all transitions and its level, useful for prediction on new-dataset.

runtime

running time in seconds of the function.

References

D. Marquardt (1963). An algorithm for least-squares estimation of nonlinear parameters. SIAM Journal of Applied Mathematics, 431-441.

See also

Author

R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>

Examples

{
if (FALSE) { # \dontrun{
library(lava)
library(prodlim)
set.seed(17)
d <- simulateIDM(n=1000)$data
fitweib <- idm(formula01=Hist(time=list(L,R),event=seen.ill)~X1+X2,
              formula02=Hist(time=observed.lifetime,event=seen.exit)~X1+X2,
              formula12=Hist(time=observed.lifetime,event=seen.exit)~X1+X2,
              data=d)
              
d <- simulateIDM(n=1000,
                 beta01=c(1,1,0,0.5,0.5,rep(0,5)),
                 beta02=c(1,0,0,0,0.5,rep(0,5)),
                 beta12=c(1,0,0,0,0.5,rep(0,5)))$data
                 
fitpenweib <- idm(formula01=Hist(time=list(L,R),
event=seen.ill)~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10,
                  formula02=Hist(time=observed.lifetime,
                  event=seen.exit)~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10,
                  formula12=~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10,
                  data=d,penalty="lasso",
                  lambda01 = c(10,20),
                  lambda02 = 10, lambda12 = 10)
} # }
}