Illness-death model algorithm with M-splines baseline risk

Illness-death model algorithm with M-splines baseline risk

idm.splines(
  b,
  clustertype,
  epsa,
  epsb,
  epsd,
  nproc,
  maxiter,
  size_V,
  size_spline,
  noVar,
  bfix,
  fix0,
  knots01,
  knots02,
  knots12,
  ctime,
  N,
  nknots01,
  nknots02,
  nknots12,
  ve01,
  ve02,
  ve12,
  dimnva01,
  dimnva02,
  dimnva12,
  nvat01,
  nvat02,
  nvat12,
  t0,
  t1,
  t2,
  t3,
  troncature,
  gausspoint,
  step.sequential,
  option.sequential
)

Arguments

b

parameters not fixed

clustertype

in which cluster to work

epsa

control convergence parameter for beta

epsb

control convergence parameter for loglik

epsd

control convergence for distance to minimum rdm

nproc

number of cluster

maxiter

Maximum number of iterations. The default is 200.

size_V

number of parameters

size_spline

number of splines parameters

noVar

indicator of variables on each transition

bfix

parameters fixed

fix0

indicators of fixed and unfixed parameters

knots01

knots of transition 0 –> 1

knots02

knots of transition 0 –> 2

knots12

knots of transition 1 –> 2

ctime

classification of subject according to their observations

N

number of subjects

nknots01

number of knots for transition 0 –>1

nknots02

number of knots for transition 0 –>2

nknots12

number of knots for transition 1 –>2

ve01

variables for transition 0 –>1

ve02

variables for transition 0 –>2

ve12

variables for transition 1 –>2

dimnva01

number of variables for transition 0 –>1

dimnva02

number of variables for transition 0 –>2

dimnva12

number of variables for transition 1 –>2

nvat01

number of variables for transition 0 –>1

nvat02

number of variables for transition 0 –>2

nvat12

number of variables for transition 1 –>2

t0

time entry

t1

time L

t2

time R

t3

time of event/out

troncature

indicator if troncature or not

gausspoint

number of points in gauss quadrature

step.sequential

should we use the optimisation version to fix splines

option.sequential

parameters to give if you want to do the optimisation version to fix splines

Author

R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>