Plot method for an illness-death model

Plot estimated baseline transition intensities from an object of class idm optionally with confidence limits.

# S3 method for class 'idm'
plot(
  x,
  lambda = NULL,
  conf.int = FALSE,
  citype = "shadow",
  add = FALSE,
  axes = TRUE,
  col,
  lwd,
  lty,
  xlim,
  ylim,
  xlab,
  ylab,
  cex,
  y.intersp,
  legend = TRUE,
  transition = c("01", "02", "12"),
  ...
)

Arguments

x

a idmWeib class object (output from calling idm with the (default) option intensities="Weib".

lambda

give the lambda for which you want to plot the penalised illness-death model

conf.int

If TRUE show confidence limits

citype

Type of confidence limits, can be "shadow" or "bars"

add

If TRUE add to existing plot

axes

If TRUE axes are drawn

col

Color of the lines

lwd

Width of the lines

lty

Type of the lines

xlim

Limits for x-axis

ylim

Limits for y-axis

xlab

Label for x-axis

ylab

Label for y-axis

cex

character expansion factor relative to current par("cex"). Used for text, and provides the default for pt.cex.

y.intersp

vertical (y) distances (in lines of text shared above/below each legend entry). A vector with one element for each row of the legend can be used.

legend

If TRUE a legend is drawn, which can be further controlled via SmartControl.

transition

Choose one of the transition intensities: c("01","02","12").

...

Passed to SmartControl

Value

Print a plot of the baseline transition intensities of an illness-death model estimated using a Weibull approach.

See also

Author

R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>

Examples


if (FALSE) { # \dontrun{
library(lava)
library(prodlim)
set.seed(17)
d <- simulateIDM(n=1000,
                 beta01=c(1,1,0,0.5,0.5,rep(0,5)),
                 beta02=c(1,0,0,0,0.5,rep(0,5)),
                 beta12=c(1,0,0,0,0.5,rep(0,5)))$data
                 
fitsplines <- idm(formula01=Hist(time=list(L,R),event=seen.ill)~X1+X2,
formula02=Hist(time=observed.lifetime,event=seen.exit)~X1+X2,
formula12=Hist(time=observed.lifetime,event=seen.exit)~X1+X2,
data=d,method="splines")
plot(fitsplines)

fitpenspline <- idm(formula01=Hist(time=list(L,R),
event=seen.ill)~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10,
                    formula02=Hist(time=observed.lifetime,
                    event=seen.exit)~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10,
                    formula12=~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10,
                    method="splines",
                    data=d,penalty="lasso",
                    lambda01 = c(10,20),
                    lambda02 = 10, lambda12 = 10)
plot(fitpenspline,lambda=c(10,10,10))
} # }