Predict transition probabilities and cumulative probabilities from an object of class idmSplines with confidence intervals are calculated.

Predictions for an illness-death model using either a penalized likelihood approach or a Weibull parametrization.

# S3 method for class 'idm'
predict(
  object,
  s,
  t,
  newdata,
  nsim = 200,
  seed = 21,
  conf.int = 0.95,
  lifeExpect = FALSE,
  maxtime,
  lambda = "BIC",
  ...
)

Arguments

object

an idm class objects returned by a call to the idm function

s

time point at which prediction is made.

t

time horizon for prediction.

newdata

A data frame with covariate values for prediction.

nsim

number of simulations for the confidence intervals calculations. The default is 200.

seed

Seed passed to set.seed for Monte Carlo simulation of confidence intervals.

conf.int

Level of confidence, i.e., a value between 0 and 1, the default is 0.95. The default is also used when conf.int=TRUE. To avoid computation of confidence intervals, set conf.int to FALSE or NULL.

lifeExpect

Logical. If TRUE compute life expectancies, i.e., t=Inf.

maxtime

The upper limit of integration for calculations of life expectancies from Weibull parametrizations.

lambda

If penalised illness-death model define if prediction need to be performed looking at the minimum of BIC or GCV or a vector of three values for each transition 0 -> 1, 0 -> 2 and 1 -> 2.

...

other parameters.

Value

a list containing the following predictions with pointwise confidence intervals:

p00

the transition probability \(p_{00}\).

p01

the transition probability \(p_{01}\).

p11

the transition probability \(p_{11}\).

p12

the transition probability \(p_{12}\).

p02_0

the probability of direct transition from state 0 to state 2.

p02_1

the probability of transition from state 0 to state 2 via state 1.

p02

transition probability \(p_{02}\). Note that p02=p_02_0+p02_1.

F01

the lifetime risk of disease. F01=p01+p02_1.

F0.

the probability of exit from state 0. F0.=p02_0+p01+p02_1.

See also

Author

R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>

Examples


if (FALSE) { # \dontrun{
library(lava)
library(prodlim)
set.seed(17)
d <- simulateIDM(n=1000,
                 beta01=c(1,1,0,0.5,0.5,rep(0,5)),
                 beta02=c(1,0,0,0,0.5,rep(0,5)),
                 beta12=c(1,0,0,0,0.5,rep(0,5)))$data
                 
fitpenweib <- idm(formula01=Hist(time=list(L,R),
event=seen.ill)~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10,
                  formula02=Hist(time=observed.lifetime,
                  event=seen.exit)~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10,
                  formula12=~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10,
                  data=d,penalty="lasso",
                  lambda01 = c(10,20),
                  lambda02 = 10, 
                  lambda12 = 10)
predict(fitpenweib,s=10,t=15,lambda = "BIC") 
} # }