Sample illness-death model data

Simulate data from an illness-death model with interval censored event times and covariates for the purpose of illustrating the help pages of the HIDeM package. See the body of the function for details, i.e., evaluate simulateIDM

simulateIDM(
  n = 100,
  seed = 1,
  scale.illtime = 2.5,
  shape.illtime = 8/100,
  scale.lifetime = 2.5,
  shape.lifetime = 8/100,
  scale.waittime = 2.5,
  shape.waittime = 8/100,
  prob.censoring = 0.05,
  administrative.censoring = 18,
  n.inspections = 8,
  schedule = 2.5,
  punctuality = 0.5,
  nvar = 10,
  mean = rep(0, 10),
  cov = matrix(c(1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1), nrow = 10, ncol = 10),
  x01 = paste0("X", 1:10),
  x02 = paste0("X", 1:10),
  x12 = paste0("X", 1:10),
  beta01 = rep(0.5, 10),
  beta02 = rep(0.5, 10),
  beta12 = rep(0.5, 10)
)

Arguments

n

number of observations

seed

specify the seed

scale.illtime

Weilbull scale for latent illness time

shape.illtime

Weilbull shape for latent illness time

scale.lifetime

Weilbull scale for latent life time

shape.lifetime

Weilbull shape for latent life time

scale.waittime

Weilbull scale for latent life time

shape.waittime

Weilbull shape for latent life time

prob.censoring

probability of censoring at each visit

administrative.censoring

specify time of administrative censoring

n.inspections

Number of inspection times

schedule

Mean of the waiting time between adjacent inspections.

punctuality

Standard deviation of waiting time between inspections.

nvar

number of variables

mean

mean of each explanatory variables

cov

covariance matrix of explanatory variables

x01

names of variables on transition 0 –> 1

x02

names of variables on transition 0 –> 2

x12

names of variables on transition 1 –> 2

beta01

value of beta on transition 0 –> 1

beta02

value of beta on transition 0 –> 2

beta12

value of beta on transition 1 –> 2

Author

R: Ariane Bercu <ariane.bercu@u-bordeaux.fr>